Only copy files if they exist

+++ heroku/0.0.1/home/entrypoint.sh
@@ -1,8 +1,13 @@
 #!/bin/sh
 
 mkdir -p $HOME/.local/share/heroku
-cp $HOME/config/config.json $HOME/.local/share/heroku/
-cp $HOME/config/.netrc $HOME
+
+if [ -f $HOME/config/config.json ]; then
+  if [ -f $HOME/config/.netrc ]; then
+    cp $HOME/config/config.json $HOME/.local/share/heroku/
+    cp $HOME/config/.netrc $HOME
+  fi
+fi
 
 #heroku $*
 /bin/sh
